% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobqueue.R
\name{any_shelf}
\alias{any_shelf}
\title{Check Shelf for Objects}
\usage{
any_shelf(x, ...)
}
\arguments{
\item{x}{a job_queue object}

\item{...}{arguments passed to methods}
}
\value{
the number of items on the shelf
}
\description{
Are there any objects on the shelf?
}
