% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_poisson.R
\name{Aquatic_Poisson}
\alias{Aquatic_Poisson}
\title{Responsible for generating mosquitoes from the aquatic stage.}
\description{
Responsible for generating mosquitoes from the aquatic stage.

Responsible for generating mosquitoes from the aquatic stage.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Aquatic_Poisson$new()}}
\item \href{#method-step}{\code{Aquatic_Poisson$step()}}
\item \href{#method-path}{\code{Aquatic_Poisson$path()}}
\item \href{#method-finalize}{\code{Aquatic_Poisson$finalize()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a set of humans
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Aquatic_Poisson$new(parameters)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameters}}{A list of parameters, including \code{lambda}, a vector of
emergence rate per site, in mosquitoes per day. Also, \code{duration} is
the number of days to run.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-step"></a>}}
\if{latex}{\out{\hypertarget{method-step}{}}}
\subsection{Method \code{step()}}{
Add a pathogen to this human's pathogens.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Aquatic_Poisson$step(step_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step_id}}{The integer ID of this step.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-path"></a>}}
\if{latex}{\out{\hypertarget{method-path}{}}}
\subsection{Method \code{path()}}{
Get the new mosquitoes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Aquatic_Poisson$path()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Destructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Aquatic_Poisson$finalize()}\if{html}{\out{</div>}}
}

}
}
