% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{build_biting_parameters}
\alias{build_biting_parameters}
\title{Make a base set of parameters for R-M mosquitoes.}
\usage{
build_biting_parameters(patch_cnt)
}
\arguments{
\item{patch_cnt}{The number of patches}

\item{year_days}{The number of days in a year}
}
\description{
Make a base set of parameters for R-M mosquitoes.
}
