% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{build_internal_parameters}
\alias{build_internal_parameters}
\title{Given parameters from a user, calculates parameters easier to compute.}
\usage{
build_internal_parameters(external_parameters)
}
\arguments{
\item{external_parameters}{The parameters as the user specified them.}

\item{adult_scale}{A scale for total mosquito counts for all patches.}
}
\value{
A list of parameters computed from the external ones so that they
are easier to use during calculation of time steps.
}
\description{
Given parameters from a user, calculates parameters easier to compute.
}
