% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{create_queue}
\alias{create_queue}
\title{Create a queue}
\usage{
create_queue(qfile, ...)
}
\arguments{
\item{qfile}{the name of the queue}

\item{...}{other arguments to be passed to \code{mdb_env}}
}
\value{
an object of class \code{"queue"} (invisibly)
fet
}
\description{
Create a new queue along with necessary files
}
\details{
The queue will be created as a subdirectory named \code{qfile}
under the current working directory
}
