% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{de_enqueue}
\alias{de_enqueue}
\title{Dequeue and Enqueue}
\usage{
de_enqueue(x, y, ...)
}
\arguments{
\item{x}{the source queue object}

\item{y}{the target queue object}

\item{...}{arguments}
}
\description{
Dequeue from one queue and add it to another queue
}
\details{
Remove the next element of \code{x} and add it to the end of \code{y}
}
