% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{dequeue}
\alias{dequeue}
\title{Pop the Next Queue Element}
\usage{
dequeue(qdb)
}
\arguments{
\item{qdb}{a queue object}
}
\value{
An R object representing the head of the queue
}
\description{
Return the next queue element and remove it from the queue
}
\details{
Return the head of the queue while also removing the element from the queue
}
