% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{enqueue}
\alias{enqueue}
\title{Push an Element}
\usage{
enqueue(qdb, val)
}
\arguments{
\item{qdb}{a queue object}

\item{val}{an R object}
}
\value{
Nothing useful
}
\description{
Push a new element on to the back of the queue
}
\details{
Insert an R object into the queue at the tail
}
