% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{extract_n}
\alias{extract_n}
\title{Extract Elements of  Queue}
\usage{
extract_n(x, ...)
}
\arguments{
\item{x}{a queue object}

\item{...}{arguments passed to other methods}
}
\value{
a list containing elements of the queue
}
\description{
Extract a Fixed Number of Elements of  Queue
}
