% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_si.R
\name{hsi_pop_events}
\alias{hsi_pop_events}
\title{Returns all events that happen at times less-than or equal-to \code{when}.
Removes those events from the queue.
When you call this, it returns the modified queue. You have to pull it out
of the return value:\preformatted{q_who <- hsi_pop_events(queue, 10.0)
queue <- q_who["queue"]
who <- q_who["chosen"]
}}
\usage{
hsi_pop_events(queue, when)
}
\description{
Returns all events that happen at times less-than or equal-to \code{when}.
Removes those events from the queue.
When you call this, it returns the modified queue. You have to pull it out
of the return value:\preformatted{q_who <- hsi_pop_events(queue, 10.0)
queue <- q_who["queue"]
who <- q_who["chosen"]
}
}
