% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_si.R
\docType{class}
\name{Human_SI}
\alias{Human_SI}
\title{Humans with susceptible-infected dynamics}
\format{
An \code{\link{R6Class}} generator object
}
\description{
Humans with susceptible-infected dynamics

Humans with susceptible-infected dynamics
}
\details{
A \code{Human_SI} is a module for MASH. It represents a set of humans in
the simulation.
}
\section{\strong{Constructor}}{

\itemize{
\item id: integer id
\item w: numeric biting weight
\item feedingID: id of the \code{\link{Feeding_Resource}} where my biting weight will be added
\item siteID: id of the \code{\link{Site}} where my feeding site resides
\item tileID: id of the \code{\link{Tile}} where I reside
}
}

\section{\strong{Methods}}{

\itemize{
\item method: i'm a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item id: integer id
\item w: numeric biting weight
\item UNBITTEN: logical flag
\item mosquito_id: integer vector of mosquitoes that have bitten me
\item mosquito_t: numeric vector of times i was bitten
\item bloodFeed: \code{FALSE} corresponds to probing-only events, \code{TRUE} corresponds to combined probe-bloodfeed events.
}
}

\keyword{R6}
\keyword{class}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Human_SI$new()}}
\item \href{#method-step}{\code{Human_SI$step()}}
\item \href{#method-infect}{\code{Human_SI$infect()}}
\item \href{#method-finalize}{\code{Human_SI$finalize()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a set of humans
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human_SI$new(parameters)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameters}}{A list of parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-step"></a>}}
\if{latex}{\out{\hypertarget{method-step}{}}}
\subsection{Method \code{step()}}{
Add a pathogen to this human's pathogens.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human_SI$step(step_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step_id}}{The integer ID of this step.
Infect an individual.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-infect"></a>}}
\if{latex}{\out{\hypertarget{method-infect}{}}}
\subsection{Method \code{infect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human_SI$infect(who, when_infected)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{who}}{An integer ID for the individual.}

\item{\code{when_infected}}{A day when the human was infected, possibly
in the past.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Destructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human_SI$finalize()}\if{html}{\out{</div>}}
}

}
}
