% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobqueue.R
\name{init_job_queue}
\alias{init_job_queue}
\title{Initialize a Job Queue}
\usage{
init_job_queue(qfile, mapsize = 2^30, ...)
}
\arguments{
\item{qfile}{the name of the job queue}

\item{mapsize}{size of the map for LMDB database}

\item{...}{other arguments to be passed to \code{mdb_env}}
}
\value{
an object of class \code{"job_queue"}
}
\description{
Initialize an existing job queue created by \code{create_job_queue}
}
\details{
\code{qfile} should be a subdirectory under the current working
directory
}
