% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{init_queue}
\alias{init_queue}
\title{Initialize a queue}
\usage{
init_queue(qfile, ...)
}
\arguments{
\item{qfile}{the name of the queue}

\item{...}{other arguments to be passed to \code{mdb_env}}
}
\value{
an object of class \code{"queue"}
}
\description{
Initialize an existing queue created by \code{create_queue}
}
\details{
\code{qfile} should be a subdirectory under the current working
directory
}
