% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobqueue.R
\name{input2shelf}
\alias{input2shelf}
\title{Move from Input to Shelf}
\usage{
input2shelf(x, ...)
}
\arguments{
\item{x}{a job_queue object}

\item{...}{other arguments passed to methods}
}
\description{
Dequeue the input queue and move that element to the shelf
}
