% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobqueue.R
\name{is_empty_output}
\alias{is_empty_output}
\title{Check if Output Queue is Empty}
\usage{
is_empty_output(x, ...)
}
\arguments{
\item{x}{a job_queue object}

\item{...}{arguments passed to other methods}
}
\value{
\code{TRUE} or \code{FALSE} depending on whether the output queue
is empty or not
}
\description{
Check to see if the output queue is empty
}
