% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{long_term}
\alias{long_term}
\title{This stochastic matrix approximates the R-M model for initial values.}
\usage{
long_term(b, parameters, location_idx)
}
\arguments{
\item{b}{The fraction that bite infectious humans each round.}

\item{parameters}{all other parameters}

\item{location_idx}{which location to model for steady-state.}
}
\value{
A long-time steady state value for M, Y, and Z as a vector.
Given emergence, EIP, and biting rate, what would be a steady-state
number of M, Y, and Z? This solves x'=Ax for steady state.
You might think we should just solve this matrix. Yes.
}
\description{
This stochastic matrix approximates the R-M model for initial values.
}
