% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\docType{class}
\name{Mosquito_RM}
\alias{Mosquito_RM}
\title{A module to represent mosquitoes at sites, as a filed of mosquitoes.}
\format{
An \code{\link{R6Class}} generator object
}
\description{
A module to represent mosquitoes at sites, as a filed of mosquitoes.

A module to represent mosquitoes at sites, as a filed of mosquitoes.
}
\details{
A \code{Mosquito_RM} is a module for MASH.
}
\keyword{R6}
\keyword{class}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Mosquito_RM$new()}}
\item \href{#method-step}{\code{Mosquito_RM$step()}}
\item \href{#method-path}{\code{Mosquito_RM$path()}}
\item \href{#method-finalize}{\code{Mosquito_RM$finalize()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a set of humans
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mosquito_RM$new(parameters)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameters}}{A list of parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-step"></a>}}
\if{latex}{\out{\hypertarget{method-step}{}}}
\subsection{Method \code{step()}}{
Add a pathogen to this human's pathogens.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mosquito_RM$step(step_id, bloodmeal_dt, aquatic)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step_id}}{The integer ID of this step.}

\item{\code{bloodmeal_dt}}{The input bloodmeal data from the last time step.
This has Bites, Time, Location in a data table.
Get mosquito density at each site.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-path"></a>}}
\if{latex}{\out{\hypertarget{method-path}{}}}
\subsection{Method \code{path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mosquito_RM$path()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Destructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mosquito_RM$finalize()}\if{html}{\out{</div>}}
}

}
}
