% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{mosquito_rm_convert_bloodmeal}
\alias{mosquito_rm_convert_bloodmeal}
\title{Convert the input bloodmeals into an internal input.}
\usage{
mosquito_rm_convert_bloodmeal(bloodmeal_dt)
}
\arguments{
\item{bloodmeal_dt}{A data table of bloodmeal inputs}
}
\value{
A matrix of kappa for the mosquito_rm model.
Dimensions are N x duration of the time step.
}
\description{
Convert the input bloodmeals into an internal input.
}
