% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{peek}
\alias{peek}
\title{Get the next of the queue}
\usage{
peek(qdb)
}
\arguments{
\item{qdb}{a queue object}
}
\value{
the value of the head of the queue
}
\description{
Return the next element of the queue
}
