% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{push.queue}
\alias{push.queue}
\title{Push item into queue}
\usage{
\method{push}{queue}(.q, v)
}
\description{
Note that it does not need to return queue obj since the environemnt is
stored like global variable
}
