% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobqueue.R
\name{shelf2input}
\alias{shelf2input}
\title{Move from Shelf to Input Queue}
\usage{
shelf2input(x, ...)
}
\arguments{
\item{x}{a job_queue object}

\item{...}{other arguments passed to methods}
}
\description{
Move all objects from the shelf to the input queue
}
