% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobqueue.R
\name{shelf2output}
\alias{shelf2output}
\title{Move from Shelf to Output Queue}
\usage{
shelf2output(x, key, val, ...)
}
\arguments{
\item{x}{a job_queue object}

\item{key}{identifier for shelf object}

\item{val}{an R object to be put in the output queue}
}
\description{
Move an object from the shelf to the output queue
}
