% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobqueue.R
\name{shelf_get}
\alias{shelf_get}
\title{Retrieve an element from the shelf}
\usage{
shelf_get(x, key, ...)
}
\arguments{
\item{x}{a job_queue object}

\item{key}{a shelf identifier key}

\item{...}{other arguments passed to methods}
}
\description{
Retrieve an element from the shelf given a key
}
