% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobqueue.R
\name{shelf_list}
\alias{shelf_list}
\title{List elements on the shelf}
\usage{
shelf_list(x, ...)
}
\arguments{
\item{x}{a job_queue object}
}
\value{
a named list of elements on the shelf
}
\description{
List elements on the shelf
}
