% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{Task}
\alias{Task}
\title{R6 class storing a task}
\description{
A task is a container that holds a function and arguments, and eventually
the results of that function when called. Various metadata fields are stored.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Task-new}{\code{Task$new()}}
\item \href{#method-Task-retrieve}{\code{Task$retrieve()}}
\item \href{#method-Task-get_task_fun}{\code{Task$get_task_fun()}}
\item \href{#method-Task-get_task_args}{\code{Task$get_task_args()}}
\item \href{#method-Task-get_task_state}{\code{Task$get_task_state()}}
\item \href{#method-Task-task_create}{\code{Task$task_create()}}
\item \href{#method-Task-task_enqueue}{\code{Task$task_enqueue()}}
\item \href{#method-Task-task_assign}{\code{Task$task_assign()}}
\item \href{#method-Task-task_start}{\code{Task$task_start()}}
\item \href{#method-Task-task_finish}{\code{Task$task_finish()}}
\item \href{#method-Task-clone}{\code{Task$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-new"></a>}}
\if{latex}{\out{\hypertarget{method-Task-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$new(fun, args = NULL, id = NULL, enqueue = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-Task-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$retrieve()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-get_task_fun"></a>}}
\if{latex}{\out{\hypertarget{method-Task-get_task_fun}{}}}
\subsection{Method \code{get_task_fun()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$get_task_fun()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-get_task_args"></a>}}
\if{latex}{\out{\hypertarget{method-Task-get_task_args}{}}}
\subsection{Method \code{get_task_args()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$get_task_args()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-get_task_state"></a>}}
\if{latex}{\out{\hypertarget{method-Task-get_task_state}{}}}
\subsection{Method \code{get_task_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$get_task_state()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-task_create"></a>}}
\if{latex}{\out{\hypertarget{method-Task-task_create}{}}}
\subsection{Method \code{task_create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$task_create()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-task_enqueue"></a>}}
\if{latex}{\out{\hypertarget{method-Task-task_enqueue}{}}}
\subsection{Method \code{task_enqueue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$task_enqueue()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-task_assign"></a>}}
\if{latex}{\out{\hypertarget{method-Task-task_assign}{}}}
\subsection{Method \code{task_assign()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$task_assign(worker_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-task_start"></a>}}
\if{latex}{\out{\hypertarget{method-Task-task_start}{}}}
\subsection{Method \code{task_start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$task_start(worker_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-task_finish"></a>}}
\if{latex}{\out{\hypertarget{method-Task-task_finish}{}}}
\subsection{Method \code{task_finish()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$task_finish(results)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Task-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
