% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_queue.R
\name{TaskQueue}
\alias{TaskQueue}
\title{R6 class for a multi-worker task queue}
\description{
Documentation baby...
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TaskQueue-new}{\code{TaskQueue$new()}}
\item \href{#method-TaskQueue-push}{\code{TaskQueue$push()}}
\item \href{#method-TaskQueue-run}{\code{TaskQueue$run()}}
\item \href{#method-TaskQueue-retrieve}{\code{TaskQueue$retrieve()}}
\item \href{#method-TaskQueue-clone}{\code{TaskQueue$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskQueue-new"></a>}}
\if{latex}{\out{\hypertarget{method-TaskQueue-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskQueue$new(workers = 4)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskQueue-push"></a>}}
\if{latex}{\out{\hypertarget{method-TaskQueue-push}{}}}
\subsection{Method \code{push()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskQueue$push(fun, args = list(), id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskQueue-run"></a>}}
\if{latex}{\out{\hypertarget{method-TaskQueue-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskQueue$run(verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskQueue-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-TaskQueue-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskQueue$retrieve()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskQueue-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaskQueue-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskQueue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
