% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_pool.R
\name{WorkerPool}
\alias{WorkerPool}
\title{R6 class storing a pool of workers}
\description{
A WorkerPool is a container that holds one or more workers and can request
them to assign tasks, start tasks, and complete tasks. It can also check to
see if any worker sessions have crashed and restart them as needed.
fields and methods that allow it to work on Tasks
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WorkerPool-new}{\code{WorkerPool$new()}}
\item \href{#method-WorkerPool-try_assign}{\code{WorkerPool$try_assign()}}
\item \href{#method-WorkerPool-try_start}{\code{WorkerPool$try_start()}}
\item \href{#method-WorkerPool-try_finish}{\code{WorkerPool$try_finish()}}
\item \href{#method-WorkerPool-refill_pool}{\code{WorkerPool$refill_pool()}}
\item \href{#method-WorkerPool-clone}{\code{WorkerPool$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkerPool-new"></a>}}
\if{latex}{\out{\hypertarget{method-WorkerPool-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerPool$new(workers = 4L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkerPool-try_assign"></a>}}
\if{latex}{\out{\hypertarget{method-WorkerPool-try_assign}{}}}
\subsection{Method \code{try_assign()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerPool$try_assign(tasks)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkerPool-try_start"></a>}}
\if{latex}{\out{\hypertarget{method-WorkerPool-try_start}{}}}
\subsection{Method \code{try_start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerPool$try_start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkerPool-try_finish"></a>}}
\if{latex}{\out{\hypertarget{method-WorkerPool-try_finish}{}}}
\subsection{Method \code{try_finish()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerPool$try_finish()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkerPool-refill_pool"></a>}}
\if{latex}{\out{\hypertarget{method-WorkerPool-refill_pool}{}}}
\subsection{Method \code{refill_pool()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerPool$refill_pool()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkerPool-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WorkerPool-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerPool$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
