% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{enqueue}
\alias{enqueue}
\title{Add an Element}
\usage{
enqueue(x, val, ...)
}
\arguments{
\item{x}{a queue object}

\item{val}{an R object}

\item{...}{arguments passed to other methods}
}
\value{
Nothing useful
}
\description{
Add a new element on to the back of the queue
}
\details{
Insert an R object into the queue at the tail
}
