% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{is_empty}
\alias{is_empty}
\title{Check if Queue is Empty}
\usage{
is_empty(x, ...)
}
\arguments{
\item{x}{a queue object}

\item{...}{arguments passed to other methods}
}
\value{
\code{TRUE} or \code{FALSE} depending on whether the queue is empty
or not
}
\description{
Check to see if the queue is empty
}
