% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{Worker}
\alias{Worker}
\title{R6 class storing a worker}
\description{
A Worker is a container that holds a callr rsession object, and possesses
fields and methods that allow it to work on Tasks
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Worker-new}{\code{Worker$new()}}
\item \href{#method-Worker-try_assign}{\code{Worker$try_assign()}}
\item \href{#method-Worker-try_start}{\code{Worker$try_start()}}
\item \href{#method-Worker-try_finish}{\code{Worker$try_finish()}}
\item \href{#method-Worker-clone}{\code{Worker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Worker-new"></a>}}
\if{latex}{\out{\hypertarget{method-Worker-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Worker$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Worker-try_assign"></a>}}
\if{latex}{\out{\hypertarget{method-Worker-try_assign}{}}}
\subsection{Method \code{try_assign()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Worker$try_assign(task)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Worker-try_start"></a>}}
\if{latex}{\out{\hypertarget{method-Worker-try_start}{}}}
\subsection{Method \code{try_start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Worker$try_start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Worker-try_finish"></a>}}
\if{latex}{\out{\hypertarget{method-Worker-try_finish}{}}}
\subsection{Method \code{try_finish()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Worker$try_finish(timeout = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Worker-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Worker-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Worker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
